V2 changes and remove V1 completely
