+++ README.md
@@ -1,111 +1,27 @@
 # PyViCare
 
-Implements an object to interact with the Viessmann ViCare API.
-The OAuth2 authentication token can optionally be stored in a file to be reused.
-Tokens are automatically renewed.
+This library implements access to Viessmann devices by using the official API from the [Viessmann Developer Portal](https://developer.viessmann.com/).
 
-A few nice feature removed from the app are available though the API (Comfort and Eco modes).
+## Breaking changes in version 1.x
 
-## Version 0.1.0
-Note that the version 0.1.0 DOES BREAK a few things.
-`ViCareSession` is now removed.
-You can now use the following objects:
-```python
-from PyViCare.PyViCareDevice import Device # generic device
-from PyViCare.PyViCareGazBoiler import GazBoiler # gaz boiler
-from PyViCare.PyViCareHeatPump import HeatPump # heat pump
-```
+* The versions prior to 1.x used an inofficial API which stopped working on July, 15th 2021. All users need to migrate to version 1.0.0 to continue using the API.
+* Python 3.4 is no longer supported.
+* Python 3.9 is now supported.
+
+## Help
+
+We need help testing and improving PyViCare, since the maintainers only have specific types of heating systems. For bugs, questions or feature requests join the [PyViCare channel on Discord](https://discord.gg/GMjBT7K5) or create an issue in this repository.
 
 ## Device Features / Errors
 
 Depending on the device, some features are not available/supported. This results in a raising of a `PyViCareNotSupportedFeatureError` if the dedicated method is called. This is most likely not a bug, but a limitation of the device itself.
-
-## Basic usage
-Simple example:
-```python
-import sys
-import logging
-sys.path.insert(0, 'PyViCare')
-from PyViCare.PyViCareGazBoiler import GazBoiler
-
-t=GazBoiler("email@domain","password","token.save")
-print(t.getDomesticHotWaterConfiguredTemperature()) 
-print(t.getDomesticHotWaterStorageTemperature())
-print(t.getOutsideTemperature())
-print(t.getRoomTemperature())
-print(t.getSupplyTemperature())
-print(t.getOutsideTemperature()) 
-print(t.getHeatingCurveShift()) 
-print(t.getHeatingCurveSlope()) 
-print(t.getBoilerTemperature())
-print(t.getActiveProgram())
-print(t.getPrograms())
-
-print(t.getCurrentDesiredTemperature())
-print(t.getMonthSinceLastService())
-print(t.getLastServiceDate())
-
-print(t.getDesiredTemperatureForProgram("comfort"))
-print(t.getActiveMode())
-
-print(t.getDesiredTemperatureForProgram("comfort"))
-print(t.setProgramTemperature("comfort",21))
-print(t.activateProgram("comfort"))
-print(t.setDomesticHotWaterTemperature(59))
-print(t.activateProgram("comfort"))
-print(t.deactivateComfort())
-```
-
-## Postman example
-
-Follow these steps to access the API in Postman:
-
-1. Create an access token in the `Authorization` tab with type `OAuth 2.0` and following inputs:
-
-    - Client id: `79742319e39245de5f91d15ff4cac2a8`
-    - Secret id: `8ad97aceb92c5892e102b093c7c083fa`
-    - Callback url: `vicare://oauth-callback/everest`
-    - Auth url: `https://iam.viessmann.com/idp/v1/authorize`
-    - Access token url: `https://iam.viessmann.com/idp/v1/token`
-    - Scope: `openid`
-
-    A login popup will open. Enter your ViCare username and password.
-
-2. Use this URL to access your `installationId` and `gatewaySerial`: 
-
-    `https://api.viessmann-platform.io/general-management/installations`
-
-    - `installationId` is `entities[0].properties.id`
-    - `gatewaySerial` is `entities[0].entities[0].properties.serial`
-
-3. Use above data to replace `{installationId}` and `{gatewaySerial}` in this URL to investigate the Viessmann API:
-
-    `https://api.viessmann-platform.io/operational-data/v1/installations/{installationId}/gateways/{gatewaySerial}/devices/0/features`
 
 ## Types of heatings
 - Use `GazBoiler` for gas heatings
 - Use `HeatPump` for heat pumps
 - Use `FuelCell` for fuel cells
 
-## Rate Limits
-
-[Due to latest changes in the Viessmann API](https://www.viessmann-community.com/t5/Konnektivitaet/Q-amp-A-Viessmann-API/td-p/127660) rate limits can be hit. In that case a `PyViCareRateLimitError` is raised. You can read from the error (`limitResetDate`) when the rate limit is reset.
-
-# Viessmann V2 API
-
-On July 15th the current Viessmann V1 API will be turned off. This library supports the Viessmann V2 API but not all properties are available (yet).
-
-We need help testing the new V2 API. Please follow these steps to test the new API.
-
-1. Register and login in the new [Viessmann Developer Portal](https://developer.viessmann.com/).
-2. In the menu navigate to `API Keys`.
-3. Create a new OAuth client using following data:
-    * Name: PyViCare
-    * Google reCAPTCHA: Disabled
-    * Redirect URIs: `vicare://oauth-callback/everest`
-4. Copy the `Client ID` to use in your code. Pass it as constructor parameter to the device.
-
-## Basic Usage with V2:
+## Basic Usage:
 
 ```python
 import sys
@@ -115,7 +31,7 @@ from PyViCare.PyViCareGazBoiler import GazBoiler
 
 client_id = "INSERT CLIENT ID"
 
-t=GazBoiler("email@domain","password","token.save", client_id=client_id, useV2=True, cacheDuration=60)
+t=GazBoiler("email@domain","password","token.save", 0, 60, client_id=client_id)
 print(t.getDomesticHotWaterConfiguredTemperature()) 
 print(t.getDomesticHotWaterStorageTemperature())
 print(t.getOutsideTemperature())
@@ -143,9 +59,9 @@ print(t.activateProgram("comfort"))
 print(t.deactivateComfort())
 ```
 
-## API V2 Usage in Postman
+## API Usage in Postman
 
-Follow these steps to access the V2 API in Postman:
+Follow these steps to access the API in Postman:
 
 1. Create an access token in the `Authorization` tab with type `OAuth 2.0` and following inputs:
 
@@ -165,18 +81,32 @@ Follow these steps to access the V2 API in Postman:
 
     A login popup will open. Enter your ViCare username and password.
 
-2. Use this URL to access your `installationId`: 
+2. Use this URL to access your `installationId` and `gatewaySerial`: 
 
-    `https://api.viessmann.com/iot/v1/equipment/installations`
+    `https://api.viessmann.com/iot/v1/equipment/installations?includeGateways=true`
 
     - `installationId` is `data[0].id`
-
-2. Use this URL to access your `gatewaySerial`. Replace `{installationId}` with the installation id above.
-
-    `https://api.viessmann.com/iot/v1/equipment/installations/{installationId}/gateways`
-
-    - `gatewaySerial` is `data[0].serial`
+    - `gatewaySerial` is `data[0].gateways[0].serial`
 
 3. Use above data to replace `{installationId}` and `{gatewaySerial}` in this URL to investigate the Viessmann API:
 
     `https://api.viessmann.com/iot/v1/equipment/installations/{installationId}/gateways/{gatewaySerial}/devices/0/features`
+
+
+## Migrate to PyViCare 1.x
+
+To use PyViCare 1.x, every user has to register and create their private API key. Follow these steps to create your API key:
+
+1. Register and login in the [Viessmann Developer Portal](https://developer.viessmann.com/).
+2. In the menu navigate to `API Keys`.
+3. Create a new OAuth client using following data:
+    * Name: PyViCare
+    * Google reCAPTCHA: Disabled
+    * Redirect URIs: `vicare://oauth-callback/everest`
+4. Copy the `Client ID` to use in your code. Pass it as constructor parameter to the device.
+
+Please not that not all previous properties are available in the new API. Missing properties were removed and might be added later if they are available again. 
+
+## Rate Limits
+
+[Due to latest changes in the Viessmann API](https://www.viessmann-community.com/t5/Konnektivitaet/Q-amp-A-Viessmann-API/td-p/127660) rate limits can be hit. In that case a `PyViCareRateLimitError` is raised. You can read from the error (`limitResetDate`) when the rate limit is reset.
